@@ -1,14 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
 
 	<PropertyGroup>
-		<TargetFrameworks>net7.0-android;net7.0-ios;net7.0-maccatalyst</TargetFrameworks>
-		<!-- ;net7.0-ios;net7.0-maccatalyst -->
-		<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows'))">$(TargetFrameworks);net7.0-windows10.0.19041.0</TargetFrameworks>
+		<TargetFrameworks>net9.0-android;net9.0-ios;net9.0-maccatalyst</TargetFrameworks>
+		<!-- ;net9.0-ios;net9.0-maccatalyst -->
+		<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows'))">$(TargetFrameworks);net9.0-windows10.0.19041.0</TargetFrameworks>
 		<!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
-		<!-- <TargetFrameworks>$(TargetFrameworks);net7.0-tizen</TargetFrameworks> -->
+		<!-- <TargetFrameworks>$(TargetFrameworks);net9.0-tizen</TargetFrameworks> -->
 		<OutputType>Exe</OutputType>
 		<RootNamespace>FinanMan.Maui</RootNamespace>
 		<UseMaui>true</UseMaui>
+		<SkipValidateMauiImplicitPackageReferences>true</SkipValidateMauiImplicitPackageReferences>
 		<SingleProject>true</SingleProject>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<EnableDefaultCssItems>false</EnableDefaultCssItems>
@@ -59,12 +60,18 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0" />
+		<ProjectReference Include="..\FinanMan.BlazorUi\FinanMan.BlazorUi.csproj" />
+		<ProjectReference Include="..\FinanMan.SharedClient\FinanMan.SharedClient.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\FinanMan.BlazorUi\FinanMan.BlazorUi.csproj" />
-		<ProjectReference Include="..\FinanMan.SharedClient\FinanMan.SharedClient.csproj" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.0" />
+		<PackageReference Update="Microsoft.AspNetCore.Components.WebView.Maui" Version="9.0.10" />
+		<PackageReference Update="Microsoft.Maui.Controls" Version="9.0.10" />
+		<PackageReference Update="Microsoft.Maui.Controls.Compatibility" Version="8.0.91" />
+	</ItemGroup>
+
+	<ItemGroup>
 	</ItemGroup>
 
 </Project>
