@@ -1,5 +1,4 @@
 ﻿using FinanMan.Shared.StateInterfaces;
-using System.Globalization;
 
 namespace FinanMan.Maui;
 
@@ -8,8 +7,7 @@ public partial class App : Application
     public App(IUiState uiState)
     {
         InitializeComponent();
-        
+
         MainPage = new MainPage(uiState);
     }
-
 }
